+++ Saitynai/Saitynai/Controllers/RegistrationsController.cs
@@ -1,9 +1,13 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
+using Saitynai.Auth.Model;
 using Saitynai.Data.Dtos.Competitions;
 using Saitynai.Data.Dtos.Events;
 using Saitynai.Data.Dtos.Registrations;
 using Saitynai.Data.Entities;
 using Saitynai.Data.Repositories;
+using System.IdentityModel.Tokens.Jwt;
+using System.Security.Claims;
 
 namespace Saitynai.Controllers
 {
@@ -14,23 +18,50 @@ namespace Saitynai.Controllers
         private readonly ICompetitionsRepository _competitionsRepository;
         private readonly IEventsRepository _eventsRepository;
         private readonly IRegistrationsRepository _registrationsRepository;
-        public RegistrationsController(IRegistrationsRepository registrationsRepository, ICompetitionsRepository competitionsRepository, IEventsRepository eventsRepository)
+        private readonly IAuthorizationService _authorizationService;
+        public RegistrationsController(IRegistrationsRepository registrationsRepository, ICompetitionsRepository competitionsRepository, IEventsRepository eventsRepository, IAuthorizationService authorizationService)
         {
             _registrationsRepository = registrationsRepository;
             _competitionsRepository = competitionsRepository;
             _eventsRepository = eventsRepository;
+            _authorizationService = authorizationService;
         }
+
+        //[HttpGet]
+        //[Authorize(Roles = IsmRoles.Admin)]
+        //public async Task<IEnumerable<RegistrationDto>> GetMany(int competitionId)
+        //{
+        //    var registrations = await _registrationsRepository.GetManyAsync(competitionId);
+
+        //    return registrations.Select(o => new RegistrationDto(o.Id, o.CarNo, o.Manufacturer, o.Model));
+        //}
 
         [HttpGet]
+        [Authorize(Roles = IsmRoles.IsmUser)]
         public async Task<IEnumerable<RegistrationDto>> GetMany(int competitionId)
         {
-            var events = await _registrationsRepository.GetManyAsync(competitionId);
+            var registrations = await _registrationsRepository.GetManyAsync(competitionId);
 
-            return events.Select(o => new RegistrationDto(o.Id, o.CarNo, o.Manufacturer, o.Model));
+            var result = new List<RegistrationDto>();
+            var i = 0;
+            foreach(var registration in registrations)
+            {
+                var authorizationResult = await _authorizationService.AuthorizeAsync(User, registration, PolicyNames.ResourceOwner);
+                if (authorizationResult.Succeeded)
+                {
+                    i++;
+                    result.Add(new RegistrationDto(registration.Id, registration.CarNo, registration.Manufacturer, registration.Model));
+                }
+            }
+
+            return result;
         }
+
+
 
         [HttpGet]
         [Route("{registrationId}", Name = "GetRegistration")]
+        [Authorize(Roles = IsmRoles.IsmUser)]
         public async Task<ActionResult<RegistrationDto>> Get(int eventId, int competitionId, int registrationId)
         {
             var eventmodel = await _eventsRepository.GetAsync(eventId);
@@ -48,10 +79,18 @@ namespace Saitynai.Controllers
             if (registration == null)
                 return NotFound();
 
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, registration, PolicyNames.ResourceOwner);
+            if (!authorizationResult.Succeeded)
+            {
+                // 404
+                return Forbid();
+            }
+
             return new RegistrationDto(registration.Id, registration.CarNo, registration.Manufacturer, registration.Model);
         }
 
         [HttpPost]
+        [Authorize(Roles = IsmRoles.IsmUser)]
         public async Task<ActionResult<RegistrationDto>> Create(int eventId, int competitionId, CreateRegistrationDto createRegistrationDto)
         {
             var eventmodel = await _eventsRepository.GetAsync(eventId);
@@ -64,7 +103,7 @@ namespace Saitynai.Controllers
             if (competition == null)
                 return NotFound();
 
-            var registration = new Registration { CarNo = createRegistrationDto.CarNo, Manufacturer = createRegistrationDto.Manufacturer, Model = createRegistrationDto.Model, Competition = competition};
+            var registration = new Registration { CarNo = createRegistrationDto.CarNo, Manufacturer = createRegistrationDto.Manufacturer, Model = createRegistrationDto.Model, Competition = competition, UserId = User.FindFirstValue(JwtRegisteredClaimNames.Sub)};
 
             await _registrationsRepository.CreateAsync(registration);
 
@@ -73,6 +112,7 @@ namespace Saitynai.Controllers
 
         [HttpPut]
         [Route("{registrationId}")]
+        [Authorize(Roles = IsmRoles.IsmUser)]
         public async Task<ActionResult<RegistrationDto>> Update(int eventId, int competitionId, int registrationId, UpdateRegistrationDto updateRegistrationDto)
         {
             var eventmodel = await _eventsRepository.GetAsync(eventId);
@@ -90,6 +130,13 @@ namespace Saitynai.Controllers
             if (registration == null)
                 return NotFound();
 
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, registration, PolicyNames.ResourceOwner);
+            if (!authorizationResult.Succeeded)
+            {
+                // 404
+                return Forbid();
+            }
+
             registration.CarNo = updateRegistrationDto.CarNo;
             registration.Manufacturer = updateRegistrationDto.Manufacturer;
             registration.Model = updateRegistrationDto.Model;
@@ -102,6 +149,7 @@ namespace Saitynai.Controllers
 
         [HttpDelete]
         [Route("{registrationId}")]
+        [Authorize(Roles = IsmRoles.IsmUser)]
         public async Task<ActionResult> Remove(int eventId, int competitionId, int registrationId)
         {
             var eventmodel = await _eventsRepository.GetAsync(eventId);
@@ -118,6 +166,13 @@ namespace Saitynai.Controllers
 
             if (registration == null)
                 return NotFound();
+
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, registration, PolicyNames.ResourceOwner);
+            if (!authorizationResult.Succeeded)
+            {
+                // 404
+                return Forbid();
+            }
 
             await _registrationsRepository.DeleteAsync(registration);
